Use sufficient large data types (S5276)

+++ include/blacklist_filter.h
@@ -48,14 +48,14 @@ namespace swd {
              *
              * @param id The id of the filter
              */
-            void set_id(const unsigned int& id);
+            void set_id(const unsigned long long& id);
 
             /**
              * @brief Get the id of the filter.
              *
              * @return The id of the filter
              */
-            unsigned int get_id() const;
+            unsigned long long get_id() const;
 
             /**
              * @brief Set the impact of the filter.
@@ -90,7 +90,7 @@ namespace swd {
             /**
              * @brief The database id of the filter.
              */
-            unsigned int id_;
+            unsigned long long id_;
 
             /**
              * @brief The impact/severity of the filter.
