+++ include/blacklist_rule.h
@@ -46,14 +46,14 @@ namespace swd {
              *
              * @param id The id of the rule
              */
-            void set_id(const unsigned int& id);
+            void set_id(const unsigned long long& id);
 
             /**
              * @brief Get the id the rule.
              *
              * @return The id of the rule
              */
-            unsigned int get_id() const;
+            unsigned long long get_id() const;
 
             /**
              * @brief Get the threshold of the rule.
@@ -73,7 +73,7 @@ namespace swd {
             /**
              * @brief The database id of the rule.
              */
-            unsigned int id_;
+            unsigned long long id_;
 
             /**
              * @brief The threshold of the rule.
